Package Readme Images
As per issue #1, images are overflowing outside of the package
readme container. Consequently, if the image width exceeds the
width of the container, the image will break out of the container
and cause the package details page to break.

In this commit I have added a max-width style (with a value of
100%) to ensure the width of the image never exceeds the width
of the `.package-readme` container.