download: Filter tasks by `status === "COMPLETED"`
Adds the task `status` to our query for the build tasks, then filter
the tasks by `status === "COMPLETED"`.

Makes sure we don't error out by attempting to look up the binary URL
for a task that was "ABORTED", "FAILED", etc.

Note: If the entire build has status "COMPLETED" (and it should, if we
got this far), then we can expect *at least one* task per OS/arch
to have status "COMPLETED".

So, in theory, filtering here should never cause us to exhaust/rule
out so much data that there are no "COMPLETED" tasks' binaries to
return the URLs of.

So, this should strictly be a no-downsides fix for the case where the
*first* run on a given OS/arch was "FAILED" or "ABORTED", etc. etc.