download: Update query string cap to 100 chars instead of 36
Neither risks much DDoS exposure, and this lets us update parameters
in the future without having to remember this limit is here and update
the limit as well.

Prevents us from accidentally breaking the microservice, although
the error messages if this got in our way would be extremely specific!

But not much downside to increasing this a bit, either.